backend/requirements.txt
# Core Backend
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
pydantic==2.10.5
pydantic-settings==2.7.1

# ML/DL Models (CPU-safe defaults)
torch==2.5.1+cpu --extra-index-url https://download.pytorch.org/whl/cpu
torchvision==0.20.1+cpu --extra-index-url https://download.pytorch.org/whl/cpu
ultralytics==8.3.64
opencv-python-headless==4.10.0.84
pillow==11.1.0
numpy==2.2.1
scikit-image==0.24.0

# Utilities
python-dotenv==1.0.1
